app/pyproject.toml
[tool.poetry]
name = "multi-modal-sample"
version = "0.1.0"
description = "LangChain + LlamaIndex + PGVector + Streamlit PDF RAG system"
authors = ["M0209 <masanorimurakoshi0209@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.1"
langchain = "^0.1.0"
langchain-core = "^0.1.46"
langchain-openai = "^0.1.6"
langchain-community = "^0.0.24"
llama_index = "^0.12.6"
llama-index-vector-stores-postgres = "^0.5.3"
tiktoken = "^0.7.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.6"
streamlit = "^1.35.0"
python-dotenv = "^1.0.1"
pdfplumber = "^0.10.3"
pdf2image = "^1.17.0"
Pillow = "^10.3.0"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
python-docx = "^1.1.0"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"